# [android] reduce JNI calls during layout measurement